implemented mariadb usage

+++ src/main/java/com/shoppinglist/backend/ItemController.java
@@ -26,11 +26,11 @@ public class ItemController {
     }
 
     @GetMapping("/items/{id}")
-    Item one(@PathVariable Long id){
+    Item one(@PathVariable int id){
         return repository.findById(id).orElseThrow(() -> new ItemNotFoundException(id));
     }
     @PutMapping("/items/{id}")
-    Item replaceItem(@RequestBody Item newItem, @PathVariable Long id){
+    Item replaceItem(@RequestBody Item newItem, @PathVariable int id){
         return repository.findById(id)
                 .map(item -> {
                     item.setName(newItem.getName());
@@ -43,7 +43,7 @@ public class ItemController {
                 });
     }
     @DeleteMapping("/items/{id}")
-    void deleteItem(@PathVariable long id){
+    void deleteItem(@PathVariable int id){
         repository.deleteById(id);
     }
 }
