+++ src/main/java/com/shoppinglist/backend/ItemNotFoundException.java
@@ -1,7 +1,7 @@
 package com.shoppinglist.backend;
 
 public class ItemNotFoundException extends RuntimeException{
-    ItemNotFoundException(long id){
+    ItemNotFoundException(int id){
         super("Could not find Item "+id);
     }
 }
