+++ src/main/java/com/shoppinglist/backend/ItemRepository.java
@@ -2,6 +2,6 @@ package com.shoppinglist.backend;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface ItemRepository extends JpaRepository<Item, Long>{
+interface ItemRepository extends JpaRepository<Item, Integer>{
 
 }
